% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asMatrixBayesLog.R
\name{as.matrix.bayesLog}
\alias{as.array.bayesLog}
\alias{as.matrix.bayesLog}
\alias{as.mcmc.bayesLog}
\title{Conversion methods for \code{bayesLog} objects}
\usage{
\method{as.matrix}{bayesLog}(x, ...)
}
\arguments{
\item{x}{a \code{bayesLog} object, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{\dots}{additional arguments to be passed to \code{\link[coda]{mcmc.convert}}}
}
\value{
A \code{matrix}, \code{array} or \code{\link[coda]{mcmc}} object
}
\description{
Methods for generic functions \code{as.matrix}, \code{as.array} and
\code{as.mcmc} for class \code{bayesLog}
}
\details{
Converts \code{\link{bayesLog}} objects to different forms.
Essentially a wrapper for \code{\link[coda]{mcmc.convert}} functions 
from the \code{coda}
package.
}
\author{
TJ McKinley
}
\seealso{
\code{\link{bayesLog}} \code{\link{plot.bayesLog}} \code{\link{window.bayesLog}} \code{\link[coda]{mcmc.convert}}
}

