% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{AUC}
\alias{AUC}
\alias{AUC.bayesLog.pred}
\alias{plot.bayesLog.auc}
\alias{print.bayesLog.auc}
\title{Generates AUC for \code{bayesLog.pred} objects}
\usage{
AUC(object, obs, ...)

\method{AUC}{bayesLog.pred}(object, obs, ...)

\method{plot}{bayesLog.auc}(x, ...)

\method{print}{bayesLog.auc}(x, ...)
}
\arguments{
\item{object}{An object of class \code{bayesLog.pred}.}

\item{obs}{A binary vector of observations.}

\item{\dots}{not used here.}

\item{x}{A \code{bayesLog.auc} object}
}
\value{
A \code{bayesLog.auc} object, essentially a vector of posterior 
samples.
}
\description{
Produce posterior for AUC from \code{bayesLog.pred} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{bayesLog.pred}: AUC method for \code{bayesLog.pred} objects

\item \code{bayesLog.auc}: Plot method for \code{bayesLog.auc} objects

\item \code{bayesLog.auc}: Print method for \code{bayesLog.auc} objects
}}

\author{
TJ McKinley
}
