% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesLog-package.R
\docType{package}
\name{BayesLog-package}
\alias{BayesLog-package}
\title{Bayesian methods for inference and model selection using logistic regression models}
\description{
Package uses MCMC methods for fitting logistic regression models in
a Bayesian framework.
}
\details{
Package provides a function \code{bayesLog} for running logistic regression models in
a Bayesian framework.
}
\author{
TJ McKinley <t.mckinley@exeter.ac.uk>
}
