% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bayesLog.R
\name{bayesLog}
\alias{bayesLog}
\title{Fits Bayesian logistic regression model}
\usage{
bayesLog(dat, response, inits = NA, inits_sigma = NA, nchains = 2,
  niter = 2e+05, scale = 0.05, varselect = F, ninitial = 10,
  priorvar = 10000, random = c("fixed", "globrand", "locrand"),
  nitertraining = NA)
}
\arguments{
\item{dat}{Data frame containing data.}

\item{response}{Character containing name of response variable.}

\item{inits}{Vector of initial values for the regression parameters.}

\item{inits_sigma}{Vector of initial values for the regression SD hyperparameters.}

\item{nchains}{Number of chains to run.}

\item{niter}{Number of iterations to run per chain.}

\item{scale}{Mixing proportion for adapt proposal.}

\item{varselect}{Logical stating whether variable selection is to be used.}

\item{ninitial}{Number of iterations before adaptive proposal kicks in.}

\item{priorvar}{Prior variance for fixed effects (currently scalar).}

\item{random}{Character: taking the values "fixed", "globrand" or "locrand" to denote fixed
effects, a global random effect or local random effects.}
}
\description{
Runs MCMC algorithm for fitting a logistic regression model in a Bayesian framework, includes options for performing variable selection.
}

