% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{confusion}
\alias{confusion}
\title{Generates confusion matrix from \code{bayesLog.pred} objects}
\usage{
confusion(x, obs, thresh, plot = T, ...)
}
\arguments{
\item{x}{An object of class \code{bayesLog.pred}.}

\item{obs}{A vector of binary observations.}

\item{thresh}{A scalar between 0 and 1 defining threshold for classification.}

\item{plot}{A logical denoting whether to plot confusion matrix or not.}

\item{\dots}{not used here.}
}
\value{
A confusion matrix, and a plot if required.
}
\description{
Produces confusion matrix from \code{\link{bayesLog.pred}} objects.
}
\author{
TJ McKinley (based on code from Mario Recker)
}

