% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crossVal.R
\name{crossVal}
\alias{crossVal}
\title{Function to perform cross-validation}
\usage{
crossVal(data, model, K = 5, plot = TRUE)
}
\arguments{
\item{data}{A data matrix.}

\item{model}{A \code{glm} model object fitted to \code{data}.}

\item{K}{Defines how many subsets to sue for cross-validation}

\item{plot}{Logical defining whether to produce plot of ROC curves.}
}
\value{
An vector containing the AUC values for the full model and the
cross-validation fits.
}
\description{
Fits logistic regression model using K-fold cross-validation
in order to explore robustness of predictions
}

