% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotBayesLog.R
\name{plot.bayesLog}
\alias{plot.bayesLog}
\title{Plots \code{bayesLog} objects}
\usage{
\method{plot}{bayesLog}(x, cond = F, trace = T, density = T, ask = T,
  randint = F, ...)
}
\arguments{
\item{x}{a \code{bayesLog} object, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{cond}{should the xs should be subsetted conditional on
inclusion in the model at each iteration.}

\item{trace}{plot trace of each variable.}

\item{density}{plot density of each variable.}

\item{ask}{prompt user before each page of plots}

\item{randint}{plot random intercepts instead of other regression variables}

\item{\dots}{additional arguments to be passed to \code{\link[coda]{plot.mcmc}} if \code{cond = FALSE}}
}
\description{
\code{plot} method for class \code{bayesLog}
}
\details{
Produces trace and/or density plots for \code{\link{bayesLog}} objects.
These are based on the \code{\link[coda]{plot.mcmc}} functions from the \code{coda}
package
}
\author{
TJ McKinley
}
\seealso{
\code{\link{bayesLog}} \code{\link{summary.bayesLog}} \code{\link{window.bayesLog}} \code{\link[coda]{plot.mcmc}}
}

