% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetBayesLog.R
\name{[.bayesLog}
\alias{[.bayesLog}
\title{Subset method for \code{bayesLog} objects}
\usage{
\method{[}{bayesLog}(x, i, j)
}
\arguments{
\item{x}{a \code{bayesLog} object, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{i}{the iterations to extract}

\item{j}{the variables to extract}

\item{\dots}{not used here}
}
\value{
A \code{bayesLog} object
}
\description{
\code{subset} method for class \code{bayesLog}
}
\details{
Extracts a subset from a \code{\link{bayesLog}} object.
Differs slightly from \code{\link[coda]{mcmc.subset}} in that it
always returns a \code{bayesLog} object, but with the iterations
reset (if \code{i} is not missing).
}
\author{
TJ McKinley
}
\seealso{
\code{\link{bayesLog}} \code{\link{plot.bayesLog}} \code{\link{window.bayesLog}}
}

