% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryBayesLog.R
\name{summary.bayesLog}
\alias{print.bayesLog}
\alias{summary.bayesLog}
\title{Summary statistics for \code{bayesLog} objects}
\usage{
\method{summary}{bayesLog}(object, vars = c("reg", "all"), ...)
}
\arguments{
\item{object}{a \code{bayesLog} object, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{vars}{a character specifying whether just the regression terms are to be
plotted (\code{"reg"}), or all of the terms (\code{"all"}).}

\item{\dots}{additional arguments to be passed to \code{\link[coda]{summary.mcmc}}}
}
\value{
A \code{summary.mcmc} object
}
\description{
\code{summary} method for class \code{bayesLog}
}
\details{
Produce summary statistics for \code{\link{bayesLog}} objects.
Essentially a wrapper for \code{\link[coda]{summary.mcmc}} functions from the \code{coda}
package.
}
\author{
TJ McKinley
}
\seealso{
\code{\link{bayesLog}} \code{\link{plot.bayesLog}} \code{\link{window.bayesLog}} \code{\link[coda]{summary.mcmc}}
}

