% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/windowBayesLog.R
\name{window.bayesLog}
\alias{window.bayesLog}
\title{Time windows for \code{bayesLog} objects}
\usage{
\method{window}{bayesLog}(x, start = NA, end = NA, thin = NA,
  chains = NA, ...)
}
\arguments{
\item{x}{a \code{bayesLog} object, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{start}{the first iteration of interest.}

\item{end}{the last iteration of interest.}

\item{thin}{the required interval between successive samples.}

\item{chains}{the chains to extract (if \code{NA} then extracts all chains).}

\item{\dots}{futher arguments for future methods.}
}
\description{
\code{summary} method for class \code{bayesLog}
}
\details{
Acts as a wrapper function for \code{\link[coda]{window.mcmc}} from the \code{coda}
package
}
\author{
TJ McKinley
}
\seealso{
\code{\link{bayesLog}} \code{\link{summary.bayesLog}} \code{\link{window.bayesLog}} \code{\link[coda]{window.mcmc}}
}

