% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auc.R
\name{AUC}
\alias{AUC}
\alias{plot.bayesLog.auc}
\alias{print.bayesLog.auc}
\title{Generates AUC for \code{bayesLog.pred} objects}
\usage{
AUC(x, ...)

\method{plot}{bayesLog.auc}(x, ...)

\method{print}{bayesLog.auc}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bayesLog.pred}.}

\item{obs}{A vector of binary observations.}

\item{\dots}{not used here.}

\item{x}{An object of class \code{bayesLog.auc}}

\item{x}{An object of class \code{bayesLog.auc}}
}
\value{
A \code{bayesLog.auc} object, essentially a vector of posterior 
samples.
}
\description{
Produce posterior for AUC from \code{\link{bayesLog.pred}} objects.
}
\section{Methods (by class)}{
\itemize{
\item \code{bayesLog.auc}: Plot method for \code{bayesLog.auc} objects.

\item \code{bayesLog.auc}: Print method for \code{bayesLog.auc} objects.
}}
\author{
TJ McKinley
}

