% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BayesLog-package.R
\docType{package}
\name{BayesLog-package}
\alias{BayesLog-package}
\title{Bayesian methods for inference and model selection using logistic regression models}
\description{
Package uses MCMC methods for fitting logistic regression models in
a Bayesian framework. Uses reversible-jump MCMC to perform variable selection
and Bayesian model averaging.
}
\details{
Package provides a function \code{bayesLog} for running logistic regression models in
a Bayesian framework. It also includes generic functions for subsetting, summarising
and plotting output.
}
\author{
TJ McKinley <t.mckinley@exeter.ac.uk>
}

