% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bayesLog.R
\name{bayesLog}
\alias{bayesLog}
\title{Fits Bayesian logistic regression model}
\usage{
bayesLog(formula, dat, gen_inits = TRUE, inits = NA, inits_rand = NA,
  priorvar = 1, prior_rand_ub = 20, nchains = 2, niter = 2e+05,
  scale = 0.05, nadapt = 100, nprintsum = 1000, maxscale = 2,
  niterdim = 1000)
}
\arguments{
\item{formula}{Formula for linear regression}

\item{dat}{Data frame containing data.}

\item{gen_inits}{Logical stating whether initial values are to be generated at random or
input by the user. If the latter then \code{inits}
must be supplied.}

\item{inits}{List containing vectors of initial values for the regression parameters.}

\item{priorvar}{A numeric specifying the prior variance for all regression terms.}

\item{nchains}{Number of chains to run.}

\item{niter}{Number of iterations to run per chain.}

\item{scale}{Mixing proportion for adapt proposal.}

\item{nadapt}{Number of iterations between each adaptive proposal update.}

\item{nprintsum}{how often to print run time information to the screen}

\item{maxscale}{the maximum scaling of the adaptive proposal variance at each update}

\item{niterdim}{the iteration at which the diminishing adaptation component kicks in}
}
\value{
An object of class \code{\link[coda]{mcmc}} or \code{\link[coda]{mcmc.list}}.
}
\description{
Runs MCMC algorithm for fitting a logistic regression model in a Bayesian framework, includes options for performing variable selection.
}

