% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesLog.R
\name{bayesLog}
\alias{bayesLog}
\title{Fits Bayesian logistic regression model}
\usage{
bayesLog(formula, dat, nsamples = NA, inits = NA, priorvar = 1,
  prior_rand_ub = 20, nchains = 2, niter = 2e+05, ninitial = 100,
  scale = 0.05, nadapt = 100, nprintsum = 1000, maxscale = 2,
  niterdim = 1000, blocks = NA, noncentreint = NA)
}
\arguments{
\item{formula}{Formula for linear regression}

\item{dat}{Data frame containing data.}

\item{nsamples}{Character specifying names of column in data to use as number of samples.}

\item{inits}{List containing vectors of initial values for the regression
parameters. If missing then these are simulated.}

\item{priorvar}{A numeric specifying the prior variance for all regression terms.}

\item{prior_rand_ub}{An upper bound on the prior standard deviation for the hierarchical terms.}

\item{nchains}{Number of chains to run.}

\item{niter}{Number of iterations to run per chain.}

\item{ninitial}{Number of iterations before adaptive proposal kicks in}

\item{scale}{Mixing proportion for adapt proposal.}

\item{nadapt}{Number of iterations between each adaptive proposal update.}

\item{nprintsum}{how often to print run time information to the screen}

\item{maxscale}{the maximum scaling of the adaptive proposal variance at each update}

\item{niterdim}{the iteration at which the diminishing adaptation component kicks in}

\item{blocks}{a list of integer vectors defining parameters to update in blocks.
Missing parameters are filled in automatically as componentwise updates.}

\item{noncentreint}{a vector of variable names for non-centring relative to the intercept.
If missing, then only centred updates performed. If \code{noncentreint == "all"} then all
updates are non-centred.}
}
\value{
An object of class \code{bayesLog}, which is basically a list
including a subset of elements:
\itemize{
\item{post:}{ an \code{\link[coda]{mcmc}} or \code{\link[coda]{mcmc.list}} containing the posterior samples.}
\item{formula:}{ a \code{formula} used to define the model.}
\item{data:}{ a \code{data.frame} containing the data used to fit the model.}
\item{nregpars:}{ a scalar containing the number of regression parameters.}
\item{nrand:}{ a scalar containing the number of random effect variables.}
\item{nrandlevels:}{ a \code{numeric} containing the number of levels of each random effect.}
}
}
\description{
Runs MCMC algorithm for fitting a logistic regression model in a Bayesian framework, includes options for performing variable selection.
}
