% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\alias{plot.bayesLog.class}
\alias{print.bayesLog.class}
\alias{summary.bayesLog.class}
\title{Generates posterior distributions for sensitivity
specificity, PPV and NPV}
\usage{
classify(pred, obs, thresh, ...)

\method{summary}{bayesLog.class}(object, ...)

\method{print}{bayesLog.class}(x, ...)

\method{plot}{bayesLog.class}(x, type = c("ind", "comp"), ...)
}
\arguments{
\item{pred}{Matrix of dimension (\code{npred} x \code{nobs}) containing predicted
values from \code{bayesLog} model.}

\item{obs}{Vector of binary observations (of length \code{nobs})}

\item{thresh}{Vector of thresholds between 0 and 1 at which to generate
classification estimates.}

\item{...}{Not currently used.}

\item{object}{An object of class \code{bayesLog.class}}

\item{x}{An object of class \code{bayesLog.class}}

\item{type}{Character vector defining whether to plot
individual or comparative plots.}
}
\value{
An object of class \code{bayesLog.class}, which is a list
including a subset of elements: \code{sens}, \code{spec}, \code{ppv}, \code{npv}
which are \code{npred} x \code{nthresh} matrices, and and element \code{thresh}
which is a vector of thresholds.
}
\description{
Generates posterior distributions for sensitivity
specificity, PPV and NPV from a set of posterior predictions.
}
\section{Methods (by generic)}{
\itemize{
\item \code{summary}: Summary method for \code{bayesLog.class} objects

\item \code{print}: Print method for \code{bayesLog.class} objects

\item \code{plot}: Plot method for \code{bayesLog.class} objects
}}

