% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotBayesLog.R
\name{plot.bayesLog}
\alias{plot.bayesLog}
\title{Plots \code{bayesLog} objects}
\usage{
\method{plot}{bayesLog}(x, vars = c("all", "reg", "rand"), rand,
  type = c("mcmc", "cat"), ...)
}
\arguments{
\item{x}{a \code{bayesLog} object, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{vars}{a character specifying whether \code{"all"} parameters are to be plotted,
or just the regression terms (\code{"reg"}), or one of the random effect terms (\code{"rand"}).
If \code{vars = "rand"}, then the user must also specify a \code{rand} argument denoting
which term to extract.}

\item{rand}{an integer denoting which random effect to extract.}

\item{type}{a character defining the type of plot to be drawn: \code{"mcmc"} denotes
trace/density plots, whereas \code{"cat"} denotes a caterpillar plot, useful for
comparing the random effects terms. (This latter plot can only be drawn for RE term currently.)}

\item{\dots}{additional arguments to be passed to \code{\link[coda]{plot.mcmc}}}
}
\description{
\code{plot} method for class \code{bayesLog}
}
\details{
Produces trace and/or density plots for \code{\link{bayesLog}} objects.
These are based on the \code{\link[coda]{plot.mcmc}} functions from the \code{coda}
package
}
\author{
TJ McKinley
}
\seealso{
\code{\link{bayesLog}} \code{\link{summary.bayesLog}} \code{\link{window.bayesLog}} \code{\link[coda]{plot.mcmc}}
}

