% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBayesLog.R
\name{plot.bayesLog}
\alias{plot.bayesLog}
\title{Plots \code{bayesLog} objects}
\usage{
plot.bayesLog(x, vars = c("reg", "all"), ...)
}
\arguments{
\item{x}{a \code{bayesLog} object, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{vars}{a character specifying whether just the regression terms are to be
plotted (\code{"reg"}), or all of the terms (\code{"all"}).}

\item{\dots}{additional arguments to be passed to \code{\link[coda]{plot.mcmc}}}
}
\description{
\code{plot} method for class \code{bayesLog}
}
\details{
Produces trace and/or density plots for \code{\link{bayesLog}} objects.
These are based on the \code{\link[coda]{plot.mcmc}} functions from the \code{coda}
package
}
\author{
TJ McKinley
}
\seealso{
\code{\link{bayesLog}} \code{\link{summary.bayesLog}} \code{\link{window.bayesLog}} 
\code{\link[coda]{plot.mcmc}}
}

