% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictBayesLog.R
\name{predict.bayesLog}
\alias{predict.bayesLog}
\title{Predict function for \code{bayesLog} objects}
\usage{
\method{predict}{bayesLog}(object, newdata, type = c("fit", "pred"), ...)
}
\arguments{
\item{object}{a \code{bayesLog} object, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{newdata}{a data frame containing the new data at which to predict}

\item{type}{a character taking the value \code{"fit"} if posterior predictive
probabilities are required, and \code{"pred"} if binary predictions are required.}

\item{\dots}{not used here}
}
\value{
A \code{bayesLog.pred} object, essentially a matrix of posterior 
predictive samples.
}
\description{
\code{predict} method for class \code{bayesLog}
}
\details{
Produce predictions for \code{\link{bayesLog}} objects.
}
\seealso{
\code{\link{bayesLog}} \code{\link[coda]{summary.mcmc}}
}
\author{
TJ McKinley
}
