% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetBayesLog.R
\name{[.bayesLog}
\alias{[.bayesLog}
\title{Subset method for \code{bayesLog} objects}
\usage{
\method{[}{bayesLog}(x, i, j)
}
\arguments{
\item{x}{a \code{bayesLog} object, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{i}{the iterations to extract}

\item{j}{the variables to extract (can be a numerical vector, or a 
character vector of variables to extract)}

\item{\dots}{not used here}
}
\value{
A \code{bayesLog} object
}
\description{
\code{subset} method for class \code{bayesLog}
}
\details{
Extracts a subset from a \code{\link{bayesLog}} object.
If extracting a single iteration, then function returns either a 
matrix or a list of matrices. Otherwise function returns a \code{bayesLog} 
object (is not subsetting by column), or an \code{mcmc} or \code{mcmc.list} 
object otherwise. If \code{i} is not missing then the iteration counts are reset.
}
\author{
TJ McKinley
}
\seealso{
\code{\link{bayesLog}} \code{\link{plot.bayesLog}} \code{\link{window.bayesLog}}
}

