% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetBayesLog.R
\name{[.bayesLog}
\alias{[.bayesLog}
\title{Subset method for \code{bayesLog} objects}
\usage{
\method{[}{bayesLog}(x, ...)
}
\arguments{
\item{x}{a \code{bayesLog} object, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{\dots}{additional arguments to be passed to \code{\link[coda]{[.mcmc}}}
}
\value{
A \code{bayesLog} object
}
\description{
\code{subset} method for class \code{bayesLog}
}
\details{
Extracts a subset from a \code{\link{bayesLog}} object.
Essentially a wrapper for \code{\link[coda]{[.mcmc}} functions from the \code{coda}
package.
}
\author{
TJ McKinley
}
\seealso{
\code{\link{bayesLog}} \code{\link{plot.bayesLog}} \code{\link{window.bayesLog}} \code{\link[coda]{[.mcmc}}
}

