% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/summaryBayesLog.R
\name{summary.bayesLog}
\alias{print.summary.bayesLog}
\alias{summary.bayesLog}
\title{Summary statistics for \code{bayesLog} objects}
\usage{
\method{summary}{bayesLog}(object, topmodels = 5, randint = F, ...)
}
\arguments{
\item{object}{a \code{bayesLog} object, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{topmodels}{the number of models to print to the screen (ordered by
posterior probability of association in decreasing order).}

\item{randint}{summarise random intercepts instead of other regression variables}

\item{\dots}{additional arguments to be passed to \code{\link[coda]{summary.mcmc}}}
}
\value{
A \code{summary.bayesLog} object; essentially a list
including a subset of elements:
\itemize{
\item{summary:}{ a \code{summary.mcmc.list} object.}
\item{PPAvar:}{ an \code{matrix} containing posterior probabilities of association
for each variable (if variable selection used).}
\item{PPAmod:}{ an \code{matrix} containing posterior probabilities of association
for each 'top' model (if variable selection used).}
}
}
\description{
\code{summary} method for class \code{bayesLog}
}
\details{
Produce summary statistics for \code{\link{bayesLog}} objects.
These are partly based on the \code{\link[coda]{summary.mcmc}} functions from the \code{coda}
package, but also include posterior probabilities of association for a set of variables or models
if variable selection was used.
}
\author{
TJ McKinley
}
\seealso{
\code{\link{bayesLog}} \code{\link{summary.bayesLog}} \code{\link{window.bayesLog}} \code{\link[coda]{summary.mcmc}}
}

