% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summaryBayesLog.R
\name{summary.bayesLog}
\alias{print.bayesLog}
\alias{summary.bayesLog}
\title{Summary statistics for \code{bayesLog} objects}
\usage{
\method{summary}{bayesLog}(object, vars = c("all", "reg", "rand"), rand, ...)
}
\arguments{
\item{object}{a \code{bayesLog} object, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{vars}{a character specifying whether \code{"all"} parameters are to be plotted,
or just the regression terms (\code{"reg"}), or one of the random effect terms (\code{"rand"}).
If \code{vars = "rand"}, then the user must also specify a \code{rand} argument denoting
which term to extract.}

\item{rand}{an integer denoting which random effect to extract.}

\item{\dots}{additional arguments to be passed to \code{\link[coda]{summary.mcmc}}}
}
\value{
A \code{summary.mcmc} object
}
\description{
\code{summary} method for class \code{bayesLog}
}
\details{
Produce summary statistics for \code{\link{bayesLog}} objects.
Essentially a wrapper for \code{\link[coda]{summary.mcmc}} functions from the \code{coda}
package.
}
\author{
TJ McKinley
}
\seealso{
\code{\link{bayesLog}} \code{\link{plot.bayesLog}} \code{\link{window.bayesLog}} \code{\link[coda]{summary.mcmc}}
}

