% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windowBayesLog.R
\name{window.bayesLog}
\alias{window.bayesLog}
\title{Time windows for \code{bayesLog} xs}
\usage{
\method{window}{bayesLog}(x, ...)
}
\arguments{
\item{x}{a \code{bayesLog} x, usually as a result of a call to
\code{\link{bayesLog}}.}

\item{\dots}{arguments to pass to \code{\link{window.mcmc}}}
}
\description{
\code{summary} method for class \code{bayesLog}
}
\details{
Acts as a wrapper function for \code{\link[coda]{window.mcmc}} from the \code{coda}
package
}
\author{
TJ McKinley
}
\seealso{
\code{\link{bayesLog}} \code{\link{summary.bayesLog}} \code{\link{window.bayesLog}} \code{\link[coda]{window.mcmc}}
}

